refactor(ModifiedAbilityModifiers): old abilityModifierRow renamed
